Crate coretime_westend_runtime

Modules

api

xcm_config

Structs

BaseDeliveryFee

The base fee for the message delivery fees.

DepositBase

One storage item; key size is 32; value is size 4+4+16+32 bytes = 56 bytes.

DepositFactor

Additional storage item size of 32 bytes.

ExistentialDeposit

FeeAssetId

The asset ID for the asset that we use to pay for message delivery fees.

FellowsBodyId

Fellows pluralistic body.

MessageQueueServiceWeight

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PotId

RelayOrigin

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SessionLength

StakingAdminBodyId

StakingAdmin pluralistic body.

TransactionByteFee

Relay Chain TransactionByteFee / 10

Version

Enums

OriginCaller

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

OFFSET

PERIOD

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Aura

AuraConfig

AuraExt

AuraExtConfig

Authorship

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CollatorSelection

CollatorSelectionConfig

CollatorSelectionUpdateOrigin

We allow Root and the StakingAdmin to execute privileged collator selection operations.

CumulusXcm

Executive

Executive: handles dispatch to the various modules.

GenesisConfigDeprecated

MessageQueue

Migrations

Migrations to apply on runtime upgrade.

Multisig

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

PriceForSiblingParachainDelivery

RootOrFellows

Privileged origin that represents Root or Fellows pluralistic body.

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

XcmpQueue